requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pymysql
